backend/requirements.txt
django>=4.0, <5.0
gunicorn==20.1.0
whitenoise
dj-database-url==1.0.0
djangorestframework
psycopg2
config==0.5.1
asgiref==3.4.1
pytz==2021.3
sqlparse==0.4.2
ccxt
